+++ src/cmd_nwr/mod.rs
@@ -13,6 +13,7 @@ pub mod label;
 pub mod lineage;
 pub mod member;
 pub mod order;
+pub mod prune;
 pub mod rename;
 pub mod replace;
 pub mod restrict;