+++ src/nwr.rs
@@ -25,6 +25,7 @@ fn main() -> anyhow::Result<()> {
         .subcommand(cmd_nwr::lineage::make_subcommand())
         .subcommand(cmd_nwr::member::make_subcommand())
         .subcommand(cmd_nwr::order::make_subcommand())
+        .subcommand(cmd_nwr::prune::make_subcommand())
         .subcommand(cmd_nwr::rename::make_subcommand())
         .subcommand(cmd_nwr::replace::make_subcommand())
         .subcommand(cmd_nwr::restrict::make_subcommand())
@@ -62,6 +63,7 @@ Subcommand groups:
         * replace
         * topo
         * subtree
+        * prune
     * Visualization
         * indent
         * comment
@@ -89,6 +91,7 @@ Subcommand groups:
         Some(("lineage", sub_matches)) => cmd_nwr::lineage::execute(sub_matches),
         Some(("member", sub_matches)) => cmd_nwr::member::execute(sub_matches),
         Some(("order", sub_matches)) => cmd_nwr::order::execute(sub_matches),
+        Some(("prune", sub_matches)) => cmd_nwr::prune::execute(sub_matches),
         Some(("rename", sub_matches)) => cmd_nwr::rename::execute(sub_matches),
         Some(("replace", sub_matches)) => cmd_nwr::replace::execute(sub_matches),
         Some(("restrict", sub_matches)) => cmd_nwr::restrict::execute(sub_matches),
@@ -110,5 +113,4 @@ Subcommand groups:
 //   nw_reroot
 // TODO:
 //   nw_match
-//   nw_prune
 //   nw_trim